update-reconect-mezon-client
