+++ packages/mezon-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "mezon-sdk",
-  "version": "2.7.80",
+  "version": "2.7.81",
   "description": "Mezon typescript SDK.",
   "files": ["dist/**"],
   "main": "dist/cjs/index.js",
